fix: update examples
