+++ docs/index.md
@@ -21,7 +21,7 @@ pip install pymmcore-widgets
     Widgets are tested on:
 
     * `macOS & Windows`
-    * `Python 3.8, 3.9 & 3.10`
+    * `Python 3.8, 3.9, 3.10 & 3.11`
     * `PyQt5 & PyQt6`
     * `PySide2 & PySide6`
