+++ examples/channel_table.py
@@ -14,7 +14,9 @@ app = QApplication([])
 mmc = CMMCorePlus().instance()
 mmc.loadSystemConfiguration()
 
-ch_table_wdg = ChannelTable()
+ch_table_wdg = ChannelTable(rows=1)
+ch_table_wdg.setChannelGroups({"Channel": ["DAPI", "FITC"]})
+ch_table_wdg.resize(500, 200)
 ch_table_wdg.show()
 
 app.exec_()
