+++ examples/grid_plan_widget.py
@@ -1,20 +1,20 @@
-"""Example usage of the ZStackWidget class.
+"""Example usage of the GridPlanWidget class.
 
-Check also the 'mda_widget.py' example to see the ZStackWidget
+Check also the 'mda_widget.py' example to see the PositionTable
 used in combination of other widgets.
 """
 
 from pymmcore_plus import CMMCorePlus
 from qtpy.QtWidgets import QApplication
 
-from pymmcore_widgets import ZStackWidget
+from pymmcore_widgets import GridPlanWidget
 
 app = QApplication([])
 
 mmc = CMMCorePlus().instance()
 mmc.loadSystemConfiguration()
 
-z_wdg = ZStackWidget()
-z_wdg.show()
+grid_wdg = GridPlanWidget()
+grid_wdg.show()
 
 app.exec_()
