--- examples/grid_widget.py
@@ -1,14 +1,0 @@
-from pymmcore_plus import CMMCorePlus
-from qtpy.QtWidgets import QApplication
-
-from pymmcore_widgets import GridWidget
-
-app = QApplication([])
-
-mmc = CMMCorePlus().instance()
-mmc.loadSystemConfiguration()
-
-grid_wdg = GridWidget()
-grid_wdg.show()
-
-app.exec_()
