+++ examples/mda_widget_useq.py
@@ -1,7 +1,7 @@
 import useq
 from qtpy.QtWidgets import QApplication
 
-from pymmcore_widgets.useq_widgets import MDASequenceWidget
+from pymmcore_widgets import MDASequenceWidget
 
 app = QApplication([])
