+++ examples/position_table.py
@@ -1,6 +1,6 @@
 """Example usage of the PositionTable class.
 
-Check also the 'mda_widget.py' example to see the PositionTable
+Check also the 'position_table.py' and 'mda_widget.py' example to see the PositionTable
 used in combination of other widgets.
 """
 
@@ -14,7 +14,8 @@ app = QApplication([])
 mmc = CMMCorePlus().instance()
 mmc.loadSystemConfiguration()
 
-pos_wdg = PositionTable()
+pos_wdg = PositionTable(rows=3)
+pos_wdg.resize(570, 200)
 pos_wdg.show()
 
 app.exec_()
