+++ examples/time_plan_widget.py
@@ -4,6 +4,7 @@ Check also the 'mda_widget.py' example to see the TimePlanWidget
 used in combination of other widgets.
 """
 
+import useq
 from pymmcore_plus import CMMCorePlus
 from qtpy.QtWidgets import QApplication
 
@@ -15,6 +16,7 @@ mmc = CMMCorePlus().instance()
 mmc.loadSystemConfiguration()
 
 t_wdg = TimePlanWidget()
+t_wdg.setValue(useq.TIntervalLoops(interval=3, loops=5))
 t_wdg.show()
 
 app.exec_()
